# api/_lib/openapi.yaml
openapi: 3.0.3
info:
  title: CoinWise AI — Custom Fintech OpenAPI
  version: 1.0.0
  description: |
    Proprietary OpenAPI server for **CoinWise AI**, the AI-driven crypto paper-trading
    & financial-intelligence platform. This server resolves the localization gaps left
    by Stripe / Plaid (which lack first-class **VND** support) and exposes endpoints
    that serve the AI **Alternative Data** analytics layer (sentiment, on-chain whale
    flow, fear & greed, AI credit score, fraud detection, AI advisor).

    Built for the Advanced Fintech Synthesis assignment — UII incubator demo.
  contact:
    name: CoinWise AI Team
    email: hello@coinwise.ai
servers:
  - url: http://localhost:3001
    description: Local dev server
...
paths:
...
  /api/v1/accounts/{accountId}/balance:
    get:
      tags: [Accounts]
      summary: Balance + holdings expressed in USD and VND
      parameters:
        - in: path
          name: accountId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Balance snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountBalance' }
  /api/v1/accounts/{accountId}/deposit-vnd:
    post:
      tags: [Accounts]
      summary: Deposit VND — auto-converted to internal USD balance
      parameters:
        - in: path
          name: accountId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amountVnd: { type: number, example: 25000000 }
                channel: { type: string, enum: [BANK_TRANSFER, MOMO, ZALOPAY, VNPAY], example: VNPAY }
              required: [amountVnd]
      responses:
        '200':
          description: Deposit confirmation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepositResponse' }
  /api/v1/accounts/{accountId}/trade:
    post:
      tags: [Accounts]
      summary: Place a paper trade (currency-safe, supports VND or USD notional)
      parameters:
        - in: path
          name: accountId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TradeRequest' }
      responses:
        '200':
          description: Trade confirmation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeResponse' }
...
components:
  schemas:
...
    FraudCheckResponse:
      type: object
      properties:
        riskScore: { type: number, minimum: 0, maximum: 1 }
        verdict: { type: string, enum: [SAFE, REVIEW, BLOCK] }
        reasons:
          type: array
          items: { type: string }
        recommendedAction: { type: string }
...
    AccountBalance:
      type: object
      properties:
        accountId: { type: string }
        cashUsd: { type: number }
        cashVnd: { type: number }
        assetsUsd: { type: number }
        assetsVnd: { type: number }
        netWorthUsd: { type: number }
        netWorthVnd: { type: number }
        positions:
          type: array
          items:
            type: object
            properties:
              symbol: { type: string }
              amount: { type: number }
              valueUsd: { type: number }
              valueVnd: { type: number }
    DepositResponse:
      type: object
      properties:
        accountId: { type: string }
        amountVnd: { type: number }
        amountUsd: { type: number }
        rate: { type: number }
        channel: { type: string }
        ref: { type: string }
        newBalanceUsd: { type: number }
    TradeRequest:
      type: object
      required: [side, symbol]
      properties:
        side: { type: string, enum: [BUY, SELL] }
        symbol: { type: string, example: BTCUSDT }
        amountUsd: { type: number }
        amountVnd: { type: number }
        amount: { type: number, description: Optional explicit base-asset amount }
    TradeResponse:
      type: object
      properties:
        ok: { type: boolean }
        txId: { type: string }
        side: { type: string }
        symbol: { type: string }
        executedAmount: { type: number }
        executedPriceUsd: { type: number }
        feeUsd: { type: number }
        newBalanceUsd: { type: number }
        fraudCheck: { $ref: '#/components/schemas/FraudCheckResponse' }
...